requirements.txt
streamlit>=1.33.0
gspread>=5.11.0
pandas>=2.2.2
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
oauthlib>=3.2.2
plotly